Luxury Items working

+++ Market_Simulation/Assets/Scripts/Economic/Jobs/ScritobleObjects/JobsOS.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "AI/Job config", fileName = "BasicJob", order = 1)]
 public class JobsOS : ScriptableObject
 {
-    public int MaxProductsCreatedInDay = 5;
-    public int MinProductsCreatedInDay = 1;
+    public float MaxProductsCreatedInDay = 5;
+    public float MinProductsCreatedInDay = 1;
 }
